requirements.txt
python-dotenv==1.0.0
DateTime==5.1
openai==0.27.2
Pillow==9.4.0
pyttsx3==2.90